Move dataset content to datasection app
